requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
pydantic==1.10.13
psycopg2-binary==2.9.9
